+++ src/components/VisibilityIcon.vue
@@ -2,10 +2,12 @@
   <icon-globe v-if="visibility === 'public'"></icon-globe>
   <icon-users v-else-if="visibility === 'private'"></icon-users>
   <icon-payment v-else-if="visibility === 'subscribers'"></icon-payment>
+  <icon-lock v-else-if="visibility === 'conversation'"></icon-lock>
   <icon-envelope v-else-if="visibility === 'direct'"></icon-envelope>
 </template>
 
 <script setup lang="ts">
+import IconLock from "@/assets/feather/lock.svg?component"
 import IconGlobe from "@/assets/feather/globe.svg?component"
 import IconEnvelope from "@/assets/forkawesome/envelope-o.svg?component"
 import IconPayment from "@/assets/tabler/coin.svg?component"