+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Support RTL text in posts and profile descriptions.
 - Allow followers-only self-replies to followers-only posts.
+- Allow replies with "conversation" visibility to followers-only posts.
 
 ### Removed